--- app/src/test/java/com/mobiledevelopmentfoodapp/com/mobiledevelopmentfoodapp/FoodTest.kt
@@ -1,11 +1,0 @@
-package com.mobiledevelopmentfoodapp.com.mobiledevelopmentfoodapp
-
-import org.junit.Assert
-import org.junit.Test
-
-class FoodTest {
-    @Test
-    fun addition_isCorrect() {
-        Assert.assertEquals(4, 2 + 2)
-    }
-}
